version 5 upgrade finished, auth error still there
