+++ routiner/src/middleware.ts
@@ -1,5 +1,5 @@
-export { auth as middleware } from "@/auth";
-export const config = { 
-    matcher: ["/board", "/calendar"],
-    runtime: 'nodejs',
-};
+import authConfig from "@/auth.config";
+import NextAuth from "next-auth";
+export const runtime = "nodejs";
+export const { auth: middleware } = NextAuth(authConfig);
+
